+++ client/i18n/en-us.json
@@ -1,0 +1,95 @@
+{
+    "ID_NO_PASS_ERROR_MSG": "password cannot be empty",
+    "ID_NO_APP_PASS_ERROR_MSG": "app password cannot be empty",
+    "ID_PASSWORD_NOT_MATCH": "passwords do not match",
+    "ID_SET_BUTTON_BUY": "Place order to buy  {{ asset }}",
+    "ID_SET_BUTTON_SELL": "Place order to sell {{ asset }}",
+    "ID_OFF": "off",
+    "ID_READY": "ready",
+    "ID_LOCKED": "locked",
+    "ID_NOWALLET": "no wallet",
+    "ID_WALLET_SYNC_PROGRESS": "wallet is {{ syncProgress }}% synced",
+    "ID_HIDE_ADDITIONAL_SETTINGS": "hide additional settings",
+    "ID_SHOW_ADDITIONAL_SETTINGS": "show additional settings",
+    "ID_BUY": "Buy",
+    "ID_SELL": "Sell",
+    "ID_NOT_SUPPORTED": "{{ asset }} is not supported",
+    "ID_CONNECTION_FAILED": "Connection to dex server failed. You can close dexc and try again later or wait for it to reconnect.",
+    "ID_ORDER_PREVIEW": "Total: {{ total }} {{ asset }}",
+    "ID_CALCULATING": "calculating...",
+    "ID_ESTIMATE_UNAVAILABLE": "estimate unavailable",
+    "ID_NO_ZERO_RATE": "zero rate not allowed",
+    "ID_NO_ZERO_QUANTITY": "zero quantity not allowed",
+    "ID_TRADE": "trade",
+    "ID_NO_ASSET_WALLET": "No {{ asset }} wallet",
+    "ID_EXECUTED": "executed",
+    "ID_BOOKED": "booked",
+    "ID_CANCELING": "canceling",
+    "ID_ACCT_UNDEFINED": "Account undefined.",
+    "ID_KEEP_WALLET_PASS": "keep current wallet password",
+    "ID_NEW_WALLET_PASS": "set a new wallet password",
+    "ID_LOT": "lot",
+    "ID_LOTS": "lots",
+    "ID_UNKNOWN": "unknown",
+    "ID_EPOCH": "epoch",
+    "ID_SETTLING": "settling",
+    "ID_NO_MATCH": "no match",
+    "ID_CANCELED": "canceled",
+    "ID_REVOKED": "revoked",
+    "ID_WAITING_FOR_CONFS": "Waiting for confirmations...",
+    "ID_NONE_SELECTED": "none selected",
+    "ID_REGISTRATION_FEE_SUCCESS": "Registration fee payment successful!",
+    "ID_API_ERROR": "API error",
+    "ID_ADD": "Add",
+    "ID_CREATE": "Create",
+    "ID_WALLET_READY": "Ready",
+    "ID_SETUP_WALLET": "Setup",
+    "ID_CHANGE_WALLET_TYPE": "change the wallet type",
+    "ID_KEEP_WALLET_TYPE": "don't change the wallet type",
+    "WALLET_READY": "Wallet Ready",
+    "SETUP_NEEDED": "Setup Needed",
+    "WALLET_PENDING": "Creating Wallet",
+    "ID_SEND_SUCCESS": "{{ assetName }} Sent!",
+    "ID_RECONFIG_SUCCESS": "Wallet Reconfigured!",
+    "ID_RESCAN_STARTED": "Wallet Rescan Running",
+    "ID_NEW_WALLET_SUCCESS": "{{ assetName }} Wallet Created!",
+    "ID_WALLET_UNLOCKED": "Wallet Unlocked",
+    "ID_SELLING": "Selling",
+    "ID_BUYING": "Buying",
+    "ID_WALLET_ENABLED_MSG": "{{ assetName }} Wallet Enabled",
+    "ID_WALLET_DISABLED_MSG": "{{ assetName }} Wallet Disabled",
+    "ID_DISABLED_MSG": "wallet is disabled",
+    "ID_ACTIVE_ORDERS_ERR_MSG": "{{ assetName }} wallet is actively managing orders",
+    "ID_AVAILABLE": "available",
+    "ID_IMMATURE": "immature",
+    "ID_FEE_BALANCE": "fee balance",
+    "ID_CANDLES_LOADING": "waiting for candlesticks",
+    "ID_DEPTH_LOADING": "retrieving depth data",
+    "ID_INVALID_ADDRESS_MSG": "invalid address: {{ address }}",
+    "ID_TXFEE_UNSUPPORTED": "fee estimation is not supported for this wallet type",
+    "ID_TXFEE_ERR_MSG": "fee estimation failed: {{ err }}",
+    "ID_ACTIVE_ORDERS_LOGOUT_ERR_MSG": "cannot logout with active orders",
+    "ID_INVALID_DATE_ERR_MSG": "error: invalid date or time",
+    "ID_NO_ARCHIVED_RECORDS": "No archived records found",
+    "ID_DELETE_ARCHIVED_RECORDS_RESULT": "Message: {{ nRecords }} archived records has been deleted",
+    "ID_ARCHIVED_RECORDS_PATH": "File Location: {{ path }}",
+    "ID_ORDER_SUBMITTING": "submitting",
+    "ID_DEFAULT": "Default",
+    "ID_ADDED": "Added",
+    "ID_DISCOVERED": "Discovered",
+    "ID_UNSUPPORTED_ASSET_INFO_ERR_MSG": "no supported asset info for id = {{ assetID }}, and no exchange info provided",
+    "ID_LIMIT_ORDER": "limit",
+    "ID_LIMIT_ORDER_IMMEDIATE_TIF": "limit (i)",
+    "ID_MARKET_ORDER": "market",
+    "ID_MATCH_STATUS_NEWLY_MATCHED": "Newly Matched",
+    "ID_MATCH_STATUS_MAKER_SWAP_CAST": "Maker Swap Sent",
+    "ID_MATCH_STATUS_TAKER_SWAP_CAST": "Taker Swap Sent",
+    "ID_MATCH_STATUS_MAKER_REDEEMED": "Maker Redeemed",
+    "ID_MATCH_STATUS_REDEMPTION_SENT": "Redemption Sent",
+    "ID_MATCH_STATUS_REVOKED": "Revoked - {{ status }}",
+    "ID_MATCH_STATUS_REFUND_PENDING": "Refund PENDING",
+    "ID_MATCH_STATUS_REFUNDED": "Refunded",
+    "ID_MATCH_STATUS_REDEEM_PENDING": "Redeem PENDING",
+    "ID_MATCH_STATUS_REDEMPTION_CONFIRMED": "Redemption Confirmed",
+    "ID_MATCH_STATUS_COMPLETE": "Complete"
+}
