+++ client/i18n/pt-br.json
@@ -1,0 +1,53 @@
+{
+    "ID_NO_PASS_ERROR_MSG": "senha não pode ser vazia",
+    "ID_NO_APP_PASS_ERROR_MSG": "senha do app não pode ser vazia",
+    "ID_PASSWORD_NOT_MATCH": "senhas diferentes",
+    "ID_SET_BUTTON_BUY": "Ordem de compra de {{ asset }}",
+    "ID_SET_BUTTON_SELL": "Ordem de venda de {{ asset }}",
+    "ID_OFF": "desligar",
+    "ID_READY": "pronto",
+    "ID_LOCKED": "trancado",
+    "ID_NOWALLET": "sem carteira",
+    "ID_WALLET_SYNC_PROGRESS": "carteira está {{ syncProgress }}% sincronizada",
+    "ID_HIDE_ADDITIONAL_SETTINGS": "esconder configurações adicionais",
+    "ID_SHOW_ADDITIONAL_SETTINGS": "mostrar configurações adicionais",
+    "ID_BUY": "Comprar",
+    "ID_SELL": "Vender",
+    "ID_NOT_SUPPORTED": "{{ asset }} não tem suporte",
+    "ID_CONNECTION_FAILED": "Conexão ao server dex falhou. Pode fechar dexc e tentar novamente depois ou esperar para tentar se reconectar.",
+    "ID_ORDER_PREVIEW": "Total: {{ total }} {{ asset }}",
+    "ID_CALCULATING": "calculando...",
+    "ID_ESTIMATE_UNAVAILABLE": "estimativa indisponível",
+    "ID_NO_ZERO_RATE": "taxa não pode ser zero",
+    "ID_NO_ZERO_QUANTITY": "quantidade não pode ser zero",
+    "ID_TRADE": "troca",
+    "ID_NO_ASSET_WALLET": "Sem carteira {{ asset }}",
+    "ID_EXECUTED": "executado",
+    "ID_BOOKED": "reservado",
+    "ID_CANCELING": "cancelando",
+    "ID_ACCT_UNDEFINED": "conta não definida.",
+    "ID_KEEP_WALLET_PASS": "manter senha da carteira",
+    "ID_NEW_WALLET_PASS": "definir nova senha para carteira",
+    "ID_LOT": "lote",
+    "ID_LOTS": "lotes",
+    "ID_UNKNOWN": "desconhecido",
+    "ID_EPOCH": "epoque",
+    "ID_SETTLING": "assentando",
+    "ID_NO_MATCH": "sem combinações",
+    "ID_CANCELED": "cancelado",
+    "ID_REVOKED": "revocado",
+    "ID_WAITING_FOR_CONFS": "Esperando confirmações...",
+    "ID_NONE_SELECTED": "nenhuma selecionado",
+    "ID_REGISTRATION_FEE_SUCCESS": "Sucesso no pagamento da taxa de registro!",
+    "ID_API_ERROR": "Erro de API",
+    "ID_ADD": "Adicionar",
+    "ID_CREATE": "Criar",
+    "ID_WALLET_READY": "Escolher",
+    "ID_SETUP_WALLET": "Configurar",
+    "ID_CHANGE_WALLET_TYPE": "trocar o tipo de carteira",
+    "ID_KEEP_WALLET_TYPE": "Não trocara tipo de carteira",
+    "WALLET_READY": "Carteira Pronta",
+    "SETUP_NEEDED": "Configuração Necessária",
+    "ID_AVAILABLE": "disponível",
+    "ID_IMMATURE": "imaturo"
+}
